requirements.txt
Django>=5.1,<6.0
gunicorn
whitenoise
djangorestframework>=3.15
psycopg2-binary>=2.9
python-decouple>=3.8
requests>=2.32
django-cors-headers>=4.3
drf-yasg>=1.21.10
drf-spectacular>=0.27.0
djangorestframework-simplejwt
django-cors-headers
python-dotenv
